done readme for GA01

+++ src/main/java/com/example/demo/repository/ActorRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.Actor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ActorRepository extends CrudRepository<Actor, Integer> {
+public interface ActorRepository extends CrudRepository<Actor, Short> {
 }
